fix(salesforce): stabilize cli utils tests
