+++ src/salesforce/cli.ts
@@ -1,302 +1,18 @@
-import * as cp from 'child_process';
-import * as os from 'os';
-import { logTrace, logWarn } from '../utils/logger';
+import { logTrace } from '../utils/logger';
 import { localize } from '../utils/localize';
 import { safeSendException } from '../shared/telemetry';
-const crossSpawn = require('cross-spawn');
 import type { OrgAuth, OrgItem } from './types';
 import * as vscode from 'vscode';
 import { CacheManager } from '../utils/cacheManager';
 import { getBooleanConfig, getNumberConfig } from '../utils/config';
-
-const CLI_TIMEOUT_MS = 120000;
-
-// Deduplicate identical execs running concurrently
-const inFlightExecs = new Map<string, Promise<{ stdout: string; stderr: string }>>();
-
-function wrapWithAbort<T>(underlying: Promise<T>, signal?: AbortSignal): Promise<T> {
-  if (!signal) {
-    return underlying;
-  }
-  if (signal.aborted) {
-    return Promise.reject(new Error('aborted'));
-  }
-  return new Promise<T>((resolve, reject) => {
-    let aborted = false;
-    const onAbort = () => {
-      aborted = true;
-      try {
-        signal.removeEventListener('abort', onAbort);
-      } catch {}
-      reject(new Error('aborted'));
-    };
-    signal.addEventListener('abort', onAbort, { once: true });
-    underlying.then(
-      v => {
-        try {
-          signal.removeEventListener('abort', onAbort);
-        } catch {}
-        if (!aborted) {
-          resolve(v);
-        }
-      },
-      err => {
-        try {
-          signal.removeEventListener('abort', onAbort);
-        } catch {}
-        if (!aborted) {
-          reject(err);
-        }
-      }
-    );
-  });
-}
-
-function makeExecKey(program: string, args: string[], envOverride?: NodeJS.ProcessEnv, timeoutMs?: number): string {
-  const hasAltPath = !!(envOverride && envOverride.PATH && envOverride.PATH !== process.env.PATH);
-  return [program, ...args, hasAltPath ? 'loginPATH' : '', String(timeoutMs || '')].join('\u0000');
-}
-
-// Allow swapping exec implementation in tests
-export type ExecFileFn = (
-  file: string,
-  args: readonly string[] | undefined,
-  options: cp.ExecFileOptionsWithStringEncoding,
-  callback: (error: NodeJS.ErrnoException | null, stdout: string, stderr: string) => void
-) => cp.ChildProcess;
-
-let execFileImpl: ExecFileFn = ((
-  file: string,
-  args: readonly string[] | undefined,
-  options: cp.ExecFileOptionsWithStringEncoding,
-  callback: (error: NodeJS.ErrnoException | null, stdout: string, stderr: string) => void
-) => {
-  const argv = Array.isArray(args) ? args.slice() : [];
-  const spawnOpts: cp.SpawnOptions = { env: options.env };
-  try {
-    logTrace('spawn:', file, argv.join(' '));
-  } catch {}
-  const child = crossSpawn(file, argv, spawnOpts);
-  let stdout = '';
-  let stderr = '';
-  const max = Math.max(1024 * 1024, options.maxBuffer || 1024 * 1024 * 10);
-  const encoding = options.encoding || 'utf8';
-  child.stdout?.setEncoding(encoding as BufferEncoding);
-  child.stderr?.setEncoding(encoding as BufferEncoding);
-  const onDataOut = (chunk: string) => {
-    stdout += chunk;
-    if (stdout.length + stderr.length > max) {
-      try {
-        child.kill();
-      } catch {}
-      const err: any = new Error('maxBuffer exceeded');
-      process.nextTick(() => callback(err, stdout, stderr));
-    }
-  };
-  const onDataErr = (chunk: string) => {
-    stderr += chunk;
-    if (stdout.length + stderr.length > max) {
-      try {
-        child.kill();
-      } catch {}
-      const err: any = new Error('maxBuffer exceeded');
-      process.nextTick(() => callback(err, stdout, stderr));
-    }
-  };
-  child.stdout?.on('data', onDataOut);
-  child.stderr?.on('data', onDataErr);
-  child.on('error', (error: any) => {
-    const err: any = error instanceof Error ? error : new Error(String(error || 'spawn error'));
-    process.nextTick(() => callback(err, stdout, stderr));
-  });
-  child.on('close', (code: number | null, signal: NodeJS.Signals | null) => {
-    if (code === 0) {
-      callback(null, stdout, stderr);
-    } else {
-      const err: any = new Error(stderr || `Command failed: ${file} ${argv.join(' ')}`);
-      if (code !== null) {
-        err.code = code;
-      }
-      if (signal) {
-        err.signal = signal;
-      }
-      callback(err, stdout, stderr);
-    }
-  });
-  return child as unknown as cp.ChildProcess;
-}) as unknown as ExecFileFn;
-
-let execOverriddenForTests = false;
-let execOverrideGeneration = 0;
-
-export function __setExecFileImplForTests(fn: ExecFileFn): void {
-  execFileImpl = fn;
-  execOverriddenForTests = true;
-  execOverrideGeneration++;
-}
-
-export function __resetExecFileImplForTests(): void {
-  execFileImpl = cp.execFile as unknown as ExecFileFn;
-  execOverriddenForTests = false;
-  execOverrideGeneration++;
-}
-
-// Lazily resolve PATH from the user's login shell (macOS/Linux) to match Terminal/Cursor
-let cachedLoginShellPATH: string | undefined;
-let resolvingPATH: Promise<string | undefined> | null = null;
-
-export async function resolvePATHFromLoginShell(): Promise<string | undefined> {
-  if (os.platform() === 'win32') {
-    return undefined;
-  }
-  if (cachedLoginShellPATH) {
-    return cachedLoginShellPATH;
-  }
-  if (resolvingPATH) {
-    return resolvingPATH;
-  }
-  resolvingPATH = new Promise<string | undefined>(resolve => {
-    const shell = process.env.SHELL || (os.platform() === 'darwin' ? '/bin/zsh' : '/bin/bash');
-    const args = ['-ilc', 'command -v printenv >/dev/null 2>&1 && printenv PATH || echo -n "$PATH"'];
-    try {
-      logTrace('resolvePATHFromLoginShell: spawn', shell, args.join(' '));
-    } catch {}
-    execFileImpl(shell, args, { maxBuffer: 1024 * 1024, encoding: 'utf8' }, (error, stdout, _stderr) => {
-      if (error) {
-        try {
-          logTrace('resolvePATHFromLoginShell: failed');
-        } catch {}
-        resolve(undefined);
-        return;
-      }
-      const pathFromShell = String(stdout || '').trim();
-      if (!pathFromShell || pathFromShell === process.env.PATH) {
-        try {
-          logTrace('resolvePATHFromLoginShell: no change');
-        } catch {}
-        resolve(undefined);
-        return;
-      }
-      cachedLoginShellPATH = pathFromShell;
-      try {
-        logTrace('resolvePATHFromLoginShell: resolved length', cachedLoginShellPATH.length);
-      } catch {}
-      resolve(cachedLoginShellPATH);
-    });
-  }).finally(() => {
-    resolvingPATH = null;
-  });
-  return resolvingPATH;
-}
-
-export async function getLoginShellEnv(): Promise<NodeJS.ProcessEnv | undefined> {
-  const loginPath = await resolvePATHFromLoginShell();
-  if (loginPath) {
-    const env2: NodeJS.ProcessEnv = { ...process.env, PATH: loginPath };
-    return env2;
-  }
-  return undefined;
-}
-
-function execCommand(
-  program: string,
-  args: string[],
-  envOverride?: NodeJS.ProcessEnv,
-  timeoutMs: number = CLI_TIMEOUT_MS,
-  signal?: AbortSignal
-): Promise<{ stdout: string; stderr: string }> {
-  const key = makeExecKey(program, args, envOverride, timeoutMs);
-  const existing = inFlightExecs.get(key);
-  if (existing) {
-    // Return a per-caller wrapper that can be cancelled without aborting the shared process
-    return wrapWithAbort(existing, signal);
-  }
-  const p = new Promise<{ stdout: string; stderr: string }>((resolve, reject) => {
-    const opts: cp.ExecFileOptionsWithStringEncoding = {
-      maxBuffer: 1024 * 1024 * 10,
-      encoding: 'utf8'
-    };
-    if (envOverride) {
-      opts.env = envOverride;
-    }
-    try {
-      logTrace('execCommand:', program, args.join(' '), envOverride?.PATH ? '(login PATH)' : '');
-    } catch {}
-    let finished = false;
-    let timer: NodeJS.Timeout;
-    const child = execFileImpl(program, args, opts, (error, stdout, stderr) => {
-      if (finished) {
-        return;
-      }
-      finished = true;
-      clearTimeout(timer);
-      inFlightExecs.delete(key);
-      if (error) {
-        const err: any = error;
-        if (err && (err.code === 'ENOENT' || /not found|ENOENT/i.test(err.message))) {
-          const cmdStr = [program, ...args].join(' ').trim();
-          const e = new Error(`CLI not found: ${cmdStr}`) as any;
-          e.code = 'ENOENT';
-          try {
-            logTrace('execCommand ENOENT for', program);
-          } catch {}
-          safeSendException('cli.exec', { code: 'ENOENT', command: program });
-          reject(e);
-          return;
-        }
-        try {
-          logTrace('execCommand error for', program, '->', (stderr || err.message || '').split('\n')[0]);
-        } catch {}
-        safeSendException('cli.exec', { code: String(err.code || ''), command: program });
-        const code = typeof err.code === 'number' || typeof err.code === 'string' ? err.code : undefined;
-        const cmdStr2 = [program, ...args].join(' ').trim();
-        const detail = stderr || err.message;
-        const msg =
-          code !== undefined
-            ? `Command "${cmdStr2}" exited with code ${code}: ${detail}`
-            : `Command "${cmdStr2}" failed: ${detail}`;
-        const e: any = new Error(msg);
-        if (code !== undefined) {
-          (e as any).code = code;
-        }
-        reject(e);
-        return;
-      }
-      try {
-        logTrace('execCommand success for', program, '(stdout length', String(stdout || '').length, ')');
-      } catch {}
-      resolve({ stdout, stderr });
-    });
-    timer = setTimeout(() => {
-      if (finished) {
-        return;
-      }
-      finished = true;
-      try {
-        child.kill();
-      } catch {}
-      try {
-        logWarn('execCommand timeout for', program, args.join(' '));
-      } catch {}
-      const cmdStrTimeout = [program, ...args].join(' ').trim();
-      const err: any = new Error(
-        localize(
-          'cliTimeout',
-          'Salesforce CLI command timed out after {0} seconds: {1}',
-          Math.round(timeoutMs / 1000),
-          cmdStrTimeout
-        )
-      );
-      err.code = 'ETIMEDOUT';
-      inFlightExecs.delete(key);
-      safeSendException('cli.exec', { code: 'ETIMEDOUT', command: program });
-      reject(err);
-    }, timeoutMs);
-  });
-  inFlightExecs.set(key, p);
-  // Per-caller cancellation should not abort the shared underlying process
-  return wrapWithAbort(p, signal);
-}
+import {
+  execCommand,
+  CLI_TIMEOUT_MS,
+  execOverriddenForTests,
+  execOverrideGeneration,
+  markExecOverriddenForTests
+} from './exec';
+import { resolvePATHFromLoginShell } from './path';
 
 // Short-lived in-memory cache for auth (avoid storing tokens on disk)
 type AuthCache = { value: OrgAuth; expiresAt: number };
@@ -499,113 +215,31 @@ export async function getOrgAuth(
   throw new Error(
     localize(
       'cliAuthFailed',
-      'Could not obtain credentials via sf/sfdx CLI. Verify authentication and try: sf org display --json --verbose'
+      'Failed to retrieve Salesforce org authentication. Run "sf org login web" to authenticate.'
     )
   );
 }
 
-function parseOrgList(json: string): OrgItem[] {
-  const parsed = JSON.parse(json);
-  const res = parsed.result || parsed;
-  let groups: any[] = [];
-  if (Array.isArray(res.orgs)) {
-    groups = groups.concat(res.orgs);
-  }
-  if (Array.isArray(res.nonScratchOrgs)) {
-    groups = groups.concat(res.nonScratchOrgs);
-  }
-  if (Array.isArray(res.scratchOrgs)) {
-    groups = groups.concat(res.scratchOrgs);
-  }
-  if (Array.isArray(res.sandboxes)) {
-    groups = groups.concat(res.sandboxes);
-  }
-  if (Array.isArray(res.devHubs)) {
-    groups = groups.concat(res.devHubs);
-  }
-  const map = new Map<string, OrgItem>();
-  for (const o of groups) {
-    const username: string | undefined = o.username || o.usernameOrAlias || o.usernameOrEmail;
-    if (!username) {
-      continue;
-    }
-    const item: OrgItem = {
-      username,
-      alias: o.alias,
-      isDefaultUsername: !!(o.isDefaultUsername || o.isDefault),
-      isDefaultDevHubUsername: !!o.isDefaultDevHubUsername,
-      isScratchOrg: !!(o.isScratchOrg || o.isScratch)
-    };
-    if (o.instanceUrl) {
-      item.instanceUrl = o.instanceUrl;
-    }
-    map.set(username, Object.assign(map.get(username) || {}, item));
-  }
-  if (Array.isArray(res.results)) {
-    for (const o of res.results) {
-      const username: string | undefined = o.username;
-      if (!username) {
-        continue;
-      }
-      const prev = map.get(username) || ({ username } as OrgItem);
-      const updated: OrgItem = {
-        ...prev,
-        alias: prev.alias || o.alias,
-        isDefaultUsername: prev.isDefaultUsername || !!o.isDefaultUsername,
-        isDefaultDevHubUsername: prev.isDefaultDevHubUsername || !!o.isDefaultDevHubUsername,
-        isScratchOrg: prev.isScratchOrg || !!o.isScratchOrg,
-        instanceUrl: prev.instanceUrl || o.instanceUrl
-      };
-      map.set(username, updated);
-    }
-  }
-  const arr = Array.from(map.values());
-  arr.sort((a, b) => {
-    const ad = a.isDefaultUsername ? 0 : 1;
-    const bd = b.isDefaultUsername ? 0 : 1;
-    if (ad !== bd) {
-      return ad - bd;
-    }
-    const an = (a.alias || a.username).toLowerCase();
-    const bn = (b.alias || b.username).toLowerCase();
-    return an.localeCompare(bn);
-  });
-  return arr;
-}
-
-// Simple in-memory cache for listOrgs
-interface OrgsCache {
-  data: OrgItem[];
-  expiresAt: number;
-  gen: number;
-}
-
-let orgsCache: OrgsCache | undefined;
-let orgsCacheTtl = 10000; // 10s default
-
-// Test hook to bypass CLI and provide deterministic results
 let listOrgsMock: (() => OrgItem[] | Promise<OrgItem[]>) | undefined;
-
-export function __setListOrgsCacheTTLForTests(ms: number): void {
-  orgsCacheTtl = ms;
-  // Avoid leaking previously cached data between tests when TTL changes
-  orgsCache = undefined;
-}
+let orgsCache: { data: OrgItem[]; expiresAt: number; gen: number } | undefined;
+let orgsCacheTtl = 60 * 1000;
 
 export function __resetListOrgsCacheForTests(): void {
   orgsCache = undefined;
-  orgsCacheTtl = 10000;
   try {
-    // Best-effort: clear persistent cache to avoid test leakage when extension is activated
     void CacheManager.delete('cli', 'orgList');
   } catch {}
   listOrgsMock = undefined;
+  orgsCacheTtl = 60 * 1000;
 }
 
 export function __setListOrgsMockForTests(fn: (() => OrgItem[] | Promise<OrgItem[]>) | undefined): void {
   listOrgsMock = fn;
-  execOverriddenForTests = true;
-  execOverrideGeneration++;
+  markExecOverriddenForTests();
+}
+
+export function __setListOrgsCacheTTLForTests(ttl: number): void {
+  orgsCacheTtl = ttl;
 }
 
 export async function listOrgs(forceRefresh = false, signal?: AbortSignal): Promise<OrgItem[]> {
@@ -724,4 +358,48 @@ export async function listOrgs(forceRefresh = false, signal?: AbortSignal): Prom
   return empty;
 }
 
-export { parseOrgList as __parseOrgListForTests, execCommand as __execCommandForTests };
+function parseOrgList(json: string): OrgItem[] {
+  const parsed = JSON.parse(json);
+  const result = parsed.result || parsed;
+  const all: OrgItem[] = [];
+  if (Array.isArray(result.orgs)) {
+    all.push(...result.orgs);
+  }
+  if (Array.isArray(result.nonScratchOrgs)) {
+    all.push(...result.nonScratchOrgs);
+  }
+  if (Array.isArray(result.scratchOrgs)) {
+    all.push(...result.scratchOrgs);
+  }
+  if (Array.isArray(result.sandboxes)) {
+    all.push(...result.sandboxes);
+  }
+  if (Array.isArray(result.devHubs)) {
+    all.push(...result.devHubs);
+  }
+  if (Array.isArray(result.results)) {
+    all.push(...result.results);
+  }
+  const dedup = new Map<string, OrgItem>();
+  for (const o of all) {
+    const key = o.username || o.alias || JSON.stringify(o);
+    if (!dedup.has(key)) {
+      dedup.set(key, o);
+    }
+  }
+  const orgs = Array.from(dedup.values());
+  orgs.sort((a, b) => {
+    if (a.isDefaultUsername && !b.isDefaultUsername) {
+      return -1;
+    }
+    if (!a.isDefaultUsername && b.isDefaultUsername) {
+      return 1;
+    }
+    const aliasA = a.alias || a.username || '';
+    const aliasB = b.alias || b.username || '';
+    return aliasA.localeCompare(aliasB);
+  });
+  return orgs;
+}
+
+export { parseOrgList as __parseOrgListForTests };
