+++ src/test/resolvePATHFromLoginShell.test.ts
@@ -1,9 +1,36 @@
 import assert from 'assert/strict';
-import { resolvePATHFromLoginShell, __setExecFileImplForTests, __resetExecFileImplForTests } from '../salesforce/cli';
+import { resolvePATHFromLoginShell, __resetLoginShellPATHForTests } from '../salesforce/path';
+import { __setExecFileImplForTests, __resetExecFileImplForTests } from '../salesforce/exec';
 
 suite('resolvePATHFromLoginShell', () => {
+  setup(() => {
+    __resetLoginShellPATHForTests();
+  });
   teardown(() => {
     __resetExecFileImplForTests();
+    __resetLoginShellPATHForTests();
+  });
+
+  test('caches PATH on success', async () => {
+    let calls = 0;
+    __setExecFileImplForTests(((program: string, args: readonly string[] | undefined, _opts: any, cb: any) => {
+      calls++;
+      cb(null, '/usr/bin', '');
+      return undefined as any;
+    }) as any);
+
+    const path1 = await resolvePATHFromLoginShell();
+    assert.equal(path1, '/usr/bin');
+
+    __setExecFileImplForTests(((program: string, args: readonly string[] | undefined, _opts: any, cb: any) => {
+      calls++;
+      cb(new Error('should not run'), '', '');
+      return undefined as any;
+    }) as any);
+
+    const path2 = await resolvePATHFromLoginShell();
+    assert.equal(path2, '/usr/bin');
+    assert.equal(calls, 1);
   });
 
   test('retries after failure', async () => {
